Move errors, kubeadm, and util under pkg.
This makes vendoring the cluster-api code into other repositories
more consistent.
