+++ clusterctl/clusterdeployer/clusterdeployer.go
@@ -23,7 +23,7 @@ import (
 	"os"
 	clustercommon "sigs.k8s.io/cluster-api/pkg/apis/cluster/common"
 	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
-	"sigs.k8s.io/cluster-api/util"
+	"sigs.k8s.io/cluster-api/pkg/util"
 	"time"
 )
