+++ pkg/controller/cluster/reconcile_test.go
@@ -25,7 +25,7 @@ import (
 	core "k8s.io/client-go/testing"
 	"sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
 	"sigs.k8s.io/cluster-api/pkg/client/clientset_generated/clientset/fake"
-	"sigs.k8s.io/cluster-api/util"
+	"sigs.k8s.io/cluster-api/pkg/util"
 )
 
 func TestClusterSetControllerReconcileHandler(t *testing.T) {
